requirements.txt
fastapi
fastapi[standard]
numpy
opencv-python
tensorflow-cpu
pillow
typing